rename inst_uncertain.* files to match convention
Closes https://github.com/physical-computation/sunflower-simulator/issues/135
